+++ app/src/main/kotlin/codes/swistak/batterymonitor/alarms/backup/AlarmBackup.kt
@@ -26,7 +26,7 @@ import java.io.InputStreamReader
 import java.nio.charset.StandardCharsets
 
 internal object AlarmBackup {
-    const val SCHEMA_VERSION: Int = Version1AlarmImporter.VERSION
+    const val SCHEMA_VERSION: Int = Version2AlarmImporter.VERSION
 
     @Throws(JSONException::class)
     fun getSchemaVersion(jsonString: String): Int = JSONObject(jsonString).optInt("version", 0)